feat(transit): stoa-rnews binary — UUCP #! rnews batch ingest
Add a stoa-rnews binary to stoa-transit that reads a #! rnews batch
from stdin (as delivered by uucico) and ingests each article into
the CID store and group log via the same check_ingest + run_pipeline
path as NNTP IHAVE.

New code:
- src/import/rnews.rs: #! rnews batch parser with gzip decompression
  (flate2); 50 MiB decompressed-size cap (UUCP-G1 gzip-bomb defense);
  1000 article/batch cap (UUCP-G3); cunbatch stubbed as unsupported
- src/rnews_config.rs: RnewsConfig — stripped config that does not
  require [listen]/[gc]/[pinning]; build_store_for_rnews delegates to
  the refactored build_store_from_parts in pipeline.rs
- src/rnews_main.rs: binary entry point; --config/--check/--help;
  exits EX_TEMPFAIL (75) on transient errors so UUCP retries;
  uses folding-aware extract_body_msgid for Message-ID extraction
- tests/rnews_ingest.rs: E2E integration test including folded
  Message-ID header case

Fixes in existing code:
- pipeline.rs: refactor build_store → build_store_from_parts so
  rnews_config.rs does not duplicate the 170-line backend match;
  fix parse_message_id_and_newsgroups early-exit bug that caused
  folded Message-ID headers to be silently stored as empty string
- ingestion.rs: make extract_body_msgid pub (needed by rnews_main
  and integration tests)
- reindex.rs: make extract_message_id_bytes pub
- Cargo.toml: add flate2 = "1", thiserror = "1"

Deferred: stoa-rmail shim, outbound uux batch sender, UucpMailer
trait — tracked in stoa-9qqyc for a follow-on issue.

+++ crates/transit/src/peering/rados_store.rs
@@ -137,7 +137,7 @@ impl IpfsStore for RadosStore {
             // Truncate to the actual bytes returned — if the object shrinks
             // between stat and read, rados_object_read returns fewer bytes.
             match ioctx.rados_object_read(&obj_name, &mut buf, 0) {
-                Ok(n) => buf.truncate(n),
+                Ok(n) => buf.truncate(n as usize),
                 Err(ceph::error::RadosError::ApiError(nix::errno::Errno::ENOENT)) => {
                     return Ok(None)
                 }
